- rename package - add instance wrapper
