+++ Example/Showcase/Classes/PresentationLayer/Chat/Chat/ChatModule.swift
@@ -1,10 +1,10 @@
 import SwiftUI
-import Chat
+import WalletConnectChat
 
 final class ChatModule {
 
     @discardableResult
-    static func create(thread: Chat.Thread, app: Application) -> UIViewController {
+    static func create(thread: WalletConnectChat.Thread, app: Application) -> UIViewController {
         let router = ChatRouter(app: app)
         let interactor = ChatInteractor(chatService: app.chatService)
         let presenter = ChatPresenter(thread: thread, interactor: interactor, router: router)
