+++ Sources/Chat/Chat.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+/// Chat instatnce wrapper
+public class Chat {
+
+    /// Chat client instance
+    public static var instance: ChatClient = {
+        return ChatClientFactory.create()
+    }()
+
+    private init() { }
+}
