+++ Sources/Chat/ChatClientFactory.swift
@@ -6,6 +6,19 @@ import WalletConnectNetworking
 
 public struct ChatClientFactory {
 
+    static func create() -> ChatClient {
+        let keychain = KeychainStorage(serviceIdentifier: "com.walletconnect.showcase")
+        let client = HTTPNetworkClient(host: "keys.walletconnect.com")
+        let registry = KeyserverRegistryProvider(client: client)
+        return ChatClientFactory.create(
+            registry: registry,
+            relayClient: Relay.instance,
+            kms: KeyManagementService(keychain: keychain),
+            logger: ConsoleLogger(),
+            keyValueStorage: UserDefaults.standard
+        )
+    }
+
     public static func create(
         registry: Registry,
         relayClient: RelayClient,
